+++ CastleVania/Screens/IntroScreen.h
@@ -7,7 +7,7 @@ class IntroScreen : public ScreenBase
 public:
 
 	void init() override;
-	void update(float dt) override;
+	void update(DWORD dt) override;
 	void renderObject() override;
 	void loadResources() override;
